Move test to reduce duplication
